Refactors the repository code to make it easier to extract a template
